Fixed warning errors caused by cooperation errors

+++ Weapon.cs
@@ -1,9 +1,9 @@
 public class Weapon{
     public int ID;
-    public double MaximumDamage;
+    public int MaximumDamage;
     public string Name;
 
-    public Weapon(int id, string name, double maxDamage){
+    public Weapon(int id, string name, int maxDamage){
         ID = id;
         Name = name;
         MaximumDamage = maxDamage;
